fixes in migration

+++ database/migrations/2021_08_30_105905_add_column_to_cash_boxes_table.php
@@ -14,7 +14,9 @@ class AddColumnToCashBoxesTable extends Migration
     public function up()
     {
         Schema::table('cash_boxes', function (Blueprint $table) {
-            $table->dropColumn('is_main');
+            if (Schema::hasColumn('cash_boxes', 'is_main')){
+                $table->dropColumn('is_main');
+            }
             $table->boolean('is_main')->nullable();
         });
     }
